+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.2.0] - unreleased
-- Override log level at runtime  `logback_bundle.core.LevelOverrideFilter`
+- Strategy based TurboFilter extension mechanism
+  - Override log level at runtime via MDC key/value
+  - Override log level per logger at runtime using dynamic strategy
 
 
 ## [0.1.0] - 2015-11-26